requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lml==0.1.0
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
openpyxl==3.0.5
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parsel==1.6.0
parso==0.7.1
pdfminer.six==20201018
pdfminer3k==1.3.4
pickleshare==0.7.5
Pillow==8.0.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.8
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyexcel==0.6.6
pyexcel-io==0.6.4
Pygments==2.7.2
PyHamcrest==2.0.2
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
queuelib==1.5.0
requests==2.25.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
sortedcontainers==2.3.0
SQLAlchemy==1.3.22
terminado==0.9.1
testpath==0.4.4
texttable==1.6.3
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
